Мелкие улучшения
- Добавлено summary для UserEntity
- Исправлено удаление категорий при загрузке по-умолчанию
- Добавлено навигационное свойство для подкатегорий
- Упрощено сидирование данных
- Переименованы методы сидирования
