+++ frontend/Money.Web/README.md
@@ -1,1 +1,2 @@
 ﻿https://www.mudblazor.com/features/icons#icons - иконки муд айкона
+https://www.mudblazor.com/utilities/spacing#how-it-works - css муд классы
